Move source files into their own directory
This change moves the source files into their own directory to better
facilitate a build process.
