+++ lib/composed.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const ObservableStore = require('..')
+const ObservableStore = require('.')
 
 
 class ComposedStore extends ObservableStore {
