+++ lib/localStorage.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const ObservableStore = require('..')
+const ObservableStore = require('.')
 
 
 class LocalStorageStore extends ObservableStore {
