+++ lib/merged.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const MergedStore = require('..')
+const MergedStore = require('.')
 
 
 class ComposedStore extends MergedStore {
